# docs/openapi.yaml
openapi: 3.0.3
info:
  title: FIGHTER HUB API
  version: 1.0.0
servers:
  - url: /api
paths:
...
  /oauth/discord/authorize:
    post:
      summary: Discordアカウント連携開始
      description: >-
        Discord OAuth Authorization Code Flow(scope=identify)を開始するためのAuthorization URLを
        発行する。Discordへの302 redirectはこのAPI自体では行わない(JWT認証が必要なため、
        ブラウザの直接navigationではAuthorizationヘッダーを付けられない)。Frontendは
        レスポンスのauthorizationUrlへ自身でnavigateする。
      security:
        - BearerAuth: []
      responses:
        '200':
          description: 発行成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DiscordAuthorizeResponse'

        '401':
          description: 認証されていない、または無効なJWT

        '500':
          description: サーバーエラー
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /oauth/discord/callback:
    get:
      summary: Discord OAuth callback
      description: >-
        Discordから呼び出されるcallback。このエンドポイントのみ認証不要(permitAll)。
        state(署名・expiration・purposeを検証)・authorization codeの検証・token交換・
        /users/@me取得・FIGHTER HUBユーザーへの連携(または付け替え)を行い、
        結果に応じてFRONTEND_BASE_URL配下の/mypageへ302 redirectする(JSONは返さない)。
        Discordのaccess/refresh tokenは保存せず、このリクエストの処理中のみ使用して破棄する。
        Discordから返されたerror_descriptionやJava例外メッセージはredirect URLへ含めず、
        固定のreasonコード(cancelled/invalid_state/expired_state/provider_error/server_error)
        のみを使用する。
      parameters:
        - name: code
          in: query
          required: false
          schema:
            type: string
          description: Discordから返された認可コード(ユーザーが認可を拒否した場合は付与されない)
        - name: state
          in: query
          required: false
          schema:
            type: string
          description: POST /api/oauth/discord/authorizeで発行した署名付きstate
        - name: error
          in: query
          required: false
          schema:
            type: string
          description: ユーザーがDiscordの認可を拒否した場合等にDiscordから付与される(例 access_denied)
      responses:
        '302':
          description: >-
            FRONTEND_BASE_URL配下の/mypageへredirectする。
            成功時: ?oauth=discord&result=success
            失敗時: ?oauth=discord&result=error&reason=<固定reason>
  /users/{id}:
    get:
      summary: ユーザー公開プロフィール取得
      description: 指定したIDのユーザーの公開プロフィールを取得する
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
          description: ユーザーID
          example: 1

      responses:
        '200':
          description: 取得成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserPublicResponse'

        '400':
          description: IDの形式が不正
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                timestamp: "2026-09-18T12:00:00"
                status: 400
                error: Bad Request
                message: "Invalid value for 'id'."
                path: /api/users/abc

        '404':
          description: 指定したユーザーが存在しない
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                timestamp: "2026-09-18T12:00:00"
                status: 404
                error: Not Found
                message: "User not found. id=999"
                path: /api/users/999

        '500':
          description: サーバーエラー
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /users/me:
    get:
      summary: 自分のプロフィール取得
      description: 認証済みユーザー自身のプロフィールを取得する
      security:
        - BearerAuth: []
      responses:
        '200':
          description: 取得成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserMeResponse'

        '401':
          description: 認証されていない、または無効なJWT

        '404':
          description: JWTのsubに対応するユーザーが存在しない
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

        '500':
          description: サーバーエラー
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    patch:
      summary: 自分のプロフィール更新
      description: 認証済みユーザー自身のプロフィールを部分更新する
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserUpdateRequest'
      responses:
        '200':
          description: 更新成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserMeResponse'

        '400':
          description: リクエストが不正、またはBean Validationエラー
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

        '401':
          description: 認証されていない、または無効なJWT

        '404':
          description: JWTのsubに対応するユーザーが存在しない
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

        '500':
          description: サーバーエラー
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /users/me/discord:
    delete:
      summary: Discordアカウント連携解除
      description: >-
        認証済みユーザー自身のDiscordアカウント連携を解除する(discord_id/discord_usernameを
        同時にNULLへ戻す)。未連携状態で呼んでも成功扱いとする冪等API。
      security:
        - BearerAuth: []
      responses:
        '204':
          description: 解除成功(未連携状態だった場合も含む)

        '401':
          description: 認証されていない、または無効なJWT

        '404':
          description: JWTのsubに対応するユーザーが存在しない
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

        '500':
          description: サーバーエラー
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /users:
    post:
      summary: ユーザー登録
      description: 新しいユーザーを登録する
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserCreateRequest'
      responses:
        '201':
          description: 登録成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserCreateResponse'

        '400':
          description: リクエストが不正、またはBean Validationエラー
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                timestamp: "2026-09-18T12:00:00"
                status: 400
                error: Bad Request
                message: "email: must be a well-formed email address"
                path: /api/users

        '409':
          description: emailが既に登録されている
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                timestamp: "2026-09-18T12:00:00"
                status: 409
                error: Conflict
                message: "Email already exists. email=user@example.com"
                path: /api/users

        '500':
          description: サーバーエラー
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  schemas:
    ErrorResponse:
      type: object
      required:
        - timestamp
        - status
        - error
        - message
        - path
      properties:
        timestamp:
          type: string
          format: date-time
          description: エラー発生日時
          example: "2026-09-18T12:00:00"
        status:
          type: integer
          description: HTTPステータスコード
          example: 400
        error:
          type: string
          description: HTTPステータスの理由句
          example: Bad Request
        message:
          type: string
          description: エラーメッセージ
          example: Invalid request
        path:
          type: string
          description: リクエストパス
          example: /api/users
...
    UserCharacterRequest:
      type: object
      required:
        - characterId
        - rank
      properties:
        characterId:
          type: integer
          format: int64
          description: キャラクターID
          example: 1
        rank:
          type: string
          description: キャラクターのランク
          example: MASTER
        mr:
          type: integer
          nullable: true
          description: Master以上の場合のMR
          example: 1800
    UserCharacterResponse:
      type: object
      required:
        - characterId
        - rank
      properties:
        characterId:
          type: integer
          format: int64
          description: キャラクターID
          example: 1
        rank:
          type: string
          description: キャラクターのランク
          example: MASTER
        mr:
          type: integer
          nullable: true
          description: Master以上の場合のMR
          example: 1800
    UserCreateRequest:
      type: object
      required:
        - email
        - password
        - name
        - characters
      properties:
        email:
          type: string
          format: email
          description: ユーザーのメールアドレス
          example: user@example.com
        password:
          type: string
          format: password
          minLength: 8
          maxLength: 100
          description: パスワード
          example: password123
        name:
          type: string
          maxLength: 50
          description: ユーザー名
          example: John Doe
        characters:
          type: array
          minItems: 1
          maxItems: 4
          description: 使用キャラクター情報(1〜4件)
          items:
            $ref: '#/components/schemas/UserCharacterRequest'
        playTimeStart:
          type: string
          format: time
          nullable: true
          description: プレイ開始時間
          example: "18:00:00"
        playTimeEnd:
          type: string
          format: time
          nullable: true
          description: プレイ終了時間
          example: "23:30:00"
        message:
          type: string
          nullable: true
          description: 自己紹介メッセージ
          example: Hello, world!
    UserCreateResponse:
      type: object
      required:
        - id
      properties:
        id:
          type: integer
          format: int64
          description: 登録されたユーザーのID
          example: 1
    DiscordAuthorizeResponse:
      type: object
      required:
        - authorizationUrl
      properties:
        authorizationUrl:
          type: string
          format: uri
          description: Frontendがnavigateすべき、Discordの認可画面へのURL(state・redirect_uri等を含む)
          example: >-
            https://discord.com/oauth2/authorize?client_id=123456789012345678&response_type=code
            &redirect_uri=http%3A%2F%2Flocalhost%3A8080%2Fapi%2Foauth%2Fdiscord%2Fcallback
            &scope=identify&state=eyJ1c2VySWQiOjF9.xxxxxxxx
    UserPublicResponse:
      type: object
      required:
        - id
        - name
        - characters
        - createdAt
        - updatedAt
      properties:
        id:
          type: integer
          format: int64
          description: ユーザーID
          example: 1
        name:
          type: string
          description: ユーザー名
          example: John Doe
        characters:
          type: array
          description: 使用キャラクター情報
          items:
            $ref: '#/components/schemas/UserCharacterResponse'
        playTimeStart:
          type: string
          format: time
          nullable: true
          description: プレイ開始時間
          example: "18:00:00"
        playTimeEnd:
          type: string
          format: time
          nullable: true
          description: プレイ終了時間
          example: "23:30:00"
        message:
          type: string
          nullable: true
          description: 自己紹介メッセージ
          example: Hello, world!
        xId:
          type: string
          nullable: true
          description: X（旧Twitter）のID
          example: johndoe123
        discordUsername:
          type: string
          nullable: true
          description: >-
            連携済みDiscordアカウントの公開用表示名(username)。内部識別子であるDiscord User ID自体は
            このAPIを含むいかなるレスポンスにも含まれない。未連携の場合はnull。
          example: johndoe
        createdAt:
          type: string
          format: date-time
          description: 作成日時
          example: "2026-09-01T00:00:00"
        updatedAt:
          type: string
          format: date-time
          description: 更新日時
          example: "2026-09-02T00:00:00"
    UserMeResponse:
      type: object
      required:
        - id
        - name
        - characters
        - createdAt
        - updatedAt
      properties:
        id:
          type: integer
          format: int64
          description: ユーザーID
          example: 1
        name:
          type: string
          description: ユーザー名
          example: John Doe
        characters:
          type: array
          description: 使用キャラクター情報
          items:
            $ref: '#/components/schemas/UserCharacterResponse'
        playTimeStart:
          type: string
          format: time
          nullable: true
          description: プレイ開始時間
          example: "18:00:00"
        playTimeEnd:
          type: string
          format: time
          nullable: true
          description: プレイ終了時間
          example: "23:30:00"
        message:
          type: string
          nullable: true
          description: 自己紹介メッセージ
          example: Hello, world!
        email:
          type: string
          format: email
          nullable: true
          description: メールアドレス(DB上NULL許容。将来的なSNSログイン対応のため)
          example: user@example.com
        xId:
          type: string
          nullable: true
          description: X（旧Twitter）のID
          example: johndoe123
        discordUsername:
          type: string
          nullable: true
          description: >-
            連携済みDiscordアカウントの公開用表示名(username)。内部識別子であるDiscord User ID自体は
            このAPIを含むいかなるレスポンスにも含まれない。未連携の場合はnull。
            連携・付け替え・解除は PATCH /api/users/me ではなく、
            POST /api/oauth/discord/authorize → GET /api/oauth/discord/callback の
            OAuthフロー、および DELETE /api/users/me/discord でのみ行う。
          example: johndoe
        createdAt:
          type: string
          format: date-time
          description: 作成日時
          example: "2026-09-01T00:00:00"
        updatedAt:
          type: string
          format: date-time
          description: 更新日時
          example: "2026-09-02T00:00:00"
    UserUpdateRequest:
      type: object
      description: >-
        部分更新用リクエスト。指定したフィールドのみ更新され、省略したフィールドは変更されない。
        name/charactersは明示的にnullを指定すると400になる。
        xId/discordIdの直接編集はこのAPIでは提供しない
        (Discordアカウント連携はPOST /api/oauth/discord/authorize経由のOAuthフローでのみ行う)。
      properties:
        name:
          type: string
          maxLength: 50
          description: ユーザー名（指定する場合は空文字不可）
          example: John Doe
        characters:
          type: array
          minItems: 1
          maxItems: 4
          description: 使用キャラクター情報(指定する場合は1〜4件)
          items:
            $ref: '#/components/schemas/UserCharacterRequest'
        playTimeStart:
          type: string
          format: time
          nullable: true
          description: プレイ開始時間
          example: "18:00:00"
        playTimeEnd:
          type: string
          format: time
          nullable: true
          description: プレイ終了時間
          example: "23:30:00"
        message:
          type: string
          nullable: true
          description: 自己紹介メッセージ
          example: Hello, world!
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT