# src/main/java/com/fighterhub/dto/UserPublicResponse.java
package com.fighterhub.dto;

import java.time.LocalDateTime;
import java.time.LocalTime;
import java.util.List;

public record UserPublicResponse(
    Long id,
    String name,
    List<UserCharacterResponse> characters,
    LocalTime playTimeStart,
    LocalTime playTimeEnd,
    String message,
    String xId,
    String discordUsername,
    LocalDateTime createdAt,
    LocalDateTime updatedAt
) {
}